pyproject.toml
[tool.poetry]
name = "ClubManager"
version = "0.1.0"
description = "An open source club management software"
authors = ["Bernard Siebens <bernard@siebens.org>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
django-constance = "^4.1.3"
django-environ = "^0.11.2"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^8.0.0" }
rules = "^3.5"
django-compressor = "^4.5.1"
pillow = "^11.0.0"
django-two-factor-auth = {extras = ["webauthn"], version = "^1.17.0"}
django-filter = "^24.3"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
flake8-django = "^1.4"
flake8-bugbear = "^24.10.31"
flake8-print = "^5.0.0"
flake8-todos = "^0.3.1"
flake8-fixme = "^1.1.1"
pytest-django = "^4.9.0"
coverage = "^7.6.4"

[tool.poetry.group.test.dependencies]
teamcity-messages = "^1.32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/migrations/*",
]

[tool.ruff]
line-length = 200
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "DJ", "T20", "TD", "FIX", "I"]
ignore = ["E111", "E114", "E117", "E501"]

[tool.black]
line-length = 150